face-recognition-app/import/python/requirements.txt
weaviate-client===3.8.0